Add api get-catalog-products Add api api-get-shopping-cart-by-user Add api-add-product-shopping-cart

+++ src/main/java/com/middleware/config/ProductCatalogConfig.java
@@ -34,7 +34,7 @@ public class ProductCatalogConfig {
 
         private String descripcion;
 
-        private float precio;
+        private double precio;
 
         @Override
         public String toString(){
